+++ src/aux-server/aux-web/aux-player/scene/MapSimulation3D.ts
@@ -35,7 +35,7 @@ import { MapPortalGrid3D, EARTH_RADIUS } from './MapPortalGrid3D';
 
 export abstract class MapSimulation3D extends PlayerSimulation3D {
     name = 'MapSimulation3D';
-    
+
     /**
      * The miniGridPortal dimension that this simulation is for.
      */