+++ src/aux-server/aux-web/aux-player/scene/PlayerMapSimulation3D.ts
@@ -23,7 +23,7 @@ import { MapSimulation3D } from './MapSimulation3D';
 
 export class PlayerMapSimulation3D extends MapSimulation3D {
     name = 'PlayerMapSimulation3D';
-    
+
     constructor(game: Game, simulation: BrowserSimulation) {
         super(MAP_PORTAL, game, simulation);
     }